+++ crates/client/block_import/src/pre_validate.rs
@@ -59,6 +59,8 @@ pub fn pre_validate_inner(
     .map(|f| f())
     .collect::<Result<(), _>>()?;
 
+    converted_classes.extend(block.trusted_converted_classes);
+
     Ok(PreValidatedBlock {
         header: block.header,
         transactions: block.transactions,
@@ -172,7 +174,17 @@ fn class_conversion(
         }
         DeclaredClass::Legacy(legacy) => {
             log::trace!("Converting legacy class with hash {:#x}", legacy.class_hash);
-            // TODO: verify that the class hash is correct
+            if !validation.trust_class_hashes {
+                let class_hash = legacy
+                    .contract_class
+                    .compute_class_hash()
+                    .map_err(|e| BlockImportError::ComputeClassHash { class_hash: legacy.class_hash, error: e })?;
+                if class_hash != legacy.class_hash {
+                    // TODO: For now we skip the exceptions for the legacy class hash mismatch
+                    log::debug!("Class hash mismatch: got {:#x}, expected {:#x}", class_hash, legacy.class_hash,);
+                    // return Err(BlockImportError::ClassHash { got: class_hash, expected: legacy.class_hash });
+                }
+            }
             Ok(ConvertedClass::Legacy(LegacyConvertedClass {
                 class_hash: legacy.class_hash,
                 info: LegacyClassInfo { contract_class: Arc::new(legacy.contract_class) },