+++ crates/node/src/service/block_production.rs
@@ -32,7 +32,7 @@ impl BlockProductionService {
         mempool: Arc<mc_mempool::Mempool>,
         block_import: Arc<BlockImporter>,
         l1_data_provider: Arc<dyn L1DataProvider>,
-        _metrics_handle: MetricsRegistry,
+        _metrics_handle: &MetricsRegistry,
         _telemetry: TelemetryHandle,
     ) -> anyhow::Result<Self> {
         if config.block_production_disabled {
@@ -66,7 +66,7 @@ impl Service for BlockProductionService {
         if is_devnet {
             // DEVNET: we the genesis block for the devnet if not deployed, otherwise we only print the devnet keys.
 
-            let keys = if (backend.get_latest_block_n().context("Getting the latest block number in db")?).is_none() {
+            let keys = if backend.get_latest_block_n().context("Getting the latest block number in db")?.is_none() {
                 // deploy devnet genesis
 
                 log::info!("⛏️  Deploying devnet genesis block");