+++ crates/node/src/service/l1.rs
@@ -27,7 +27,7 @@ impl L1SyncService {
     pub async fn new(
         config: &L1SyncParams,
         db: &DatabaseService,
-        metrics_handle: MetricsRegistry,
+        metrics_handle: &MetricsRegistry,
         l1_gas_provider: GasPriceProvider,
         chain_id: ChainId,
         l1_core_address: H160,
@@ -37,7 +37,7 @@ impl L1SyncService {
             if let Some(l1_rpc_url) = &config.l1_endpoint {
                 let core_address = Address::from_slice(l1_core_address.as_bytes());
                 let l1_block_metrics =
-                    L1BlockMetrics::register(&metrics_handle).expect("Registering prometheus metrics");
+                    L1BlockMetrics::register(metrics_handle).expect("Registering prometheus metrics");
                 Some(
                     EthereumClient::new(l1_rpc_url.clone(), core_address, l1_block_metrics)
                         .await