+++ crates/primitives/convert/src/lib.rs
@@ -2,7 +2,7 @@ mod felt;
 mod to_felt;
 
 pub use felt::{felt_to_u128, felt_to_u64};
-pub use to_felt::ToFelt;
+pub use to_felt::{DisplayFeltAsHex, FeltHexDisplay, ToFelt};
 
 pub mod test {
     /// Asserts that the conversion between two types is consistent.